prioritizrdata: Conservation Planning Datasets

The prioritizrdata package is a supplemental package that contains example datasets for conservation planning. It is intended to be used alongside the prioritizr package.

Details

This package contains the following datasets:

tas_data

Conservation planning dataset for Tasmania, Australia.

salt_data

Conservation planning dataset for Salt Spring Island, Canada.

wa_data

Conservation planning dataset for Washington, The United States of America.

See also

Author

Maintainer: Richard Schuster richard.schuster@glel.carleton.ca (ORCID)

Authors:

Jeffrey O Hanson jeffrey.hanson@uqconnect.edu.au (ORCID)

Nina Morrell nina.morrell@ubc.ca

Matthew Strimas-Mackey mstrimas@gmail.com (ORCID)

Kristin Camille V Buenafe k.buenafe@uq.edu.au (ORCID)

Matthew E Watts m.watts@uq.edu.au

Peter Arcese peter.arcese@ubc.ca (ORCID)

Joseph R Bennett joseph.bennett@carleton.ca (ORCID)

Hugh P Possingham hugh.possingham@tnc.org (ORCID)